% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_base_req}
\alias{create_base_req}
\title{Create base url for Omeka S installation
The API of a Omeka S installation can be found at the URL `https://youromekeas.domain/api`.
To access (private) data you need a `key_identity` and `key_credential`, you get both in the Omeka S backend.}
\usage{
create_base_req(omeka_url, key_identity, key_credential)
}
\arguments{
\item{omeka_url}{The URL of the Omeka S installation}

\item{key_identity}{your `key_identity`, preferably saved in `.Renviron`}

\item{key_credential}{your `key_credential`, preferably saved in `.Renviron`}
}
\description{
Create base url for Omeka S installation
The API of a Omeka S installation can be found at the URL `https://youromekeas.domain/api`.
To access (private) data you need a `key_identity` and `key_credential`, you get both in the Omeka S backend.
}
