% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_omekas.R
\name{query_items_by_id}
\alias{query_items_by_id}
\title{Query items by resource class id}
\usage{
query_items_by_id(
  base_req,
  per_page = 99999,
  query_string = "items",
  resource_class_id = NULL
)
}
\arguments{
\item{base_req}{base request, typically result of `create_base_request`}

\item{per_page}{integer, number of items to fetch, default set to `99999`}

\item{query_string}{fetch data from `items` endpoint}

\item{resource_class_id}{ressource_id}
}
\value{
items with this vocabulary id
}
\description{
Query items by resource class id
}
\examples{
\dontrun{
query_items_by_vocab_id(base_req, ressource_id = 1)
}
}
