% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_omekas.R
\name{query_ressource_classes}
\alias{query_ressource_classes}
\title{Query ressource classes}
\usage{
query_ressource_classes(
  base_req,
  per_page = 99999,
  query_string = "resource_classes",
  get_all = NULL,
  vocabulary_prefix = NULL,
  term = NULL
)
}
\arguments{
\item{base_req}{base request, typically result of `create_base_request`}

\item{per_page}{integer, number of items to fetch, default set to `99999`}

\item{query_string}{in this case harded coded to `resource_classes`}

\item{get_all}{boolean: if `TRUE` (or not null, to be precise) all `ressource_classes` will be fetched}

\item{vocabulary_prefix}{string, get all classes for special vocabulary, e.g. `dcterms`, `foaf`}

\item{term}{string, search for terms, e.g. `dcterms:title`, `foaf:Person`}
}
\value{
the results from the search
}
\description{
This function get data from the `resource_classes` endpoint.
}
\examples{
\dontrun{
query_ressource_classes(req, vocabulary_prefix = "dcterms")
query_ressource_classes(req, get_all = TRUE)
}
}
